refactor: hoist db/ test helpers into test_support (#306)
Centralize the cross-cutting helpers the db unit tests previously
duplicated or scattered across `*::test_helpers` submodules. Per the
**Shared helpers** rule in `.claude/rules/03-unit-testing.md`, every
crate now gates a `test_support` module behind
`#[cfg(any(test, feature = "test-support"))]` as the single source of
truth.

Relocated to `db/src/test_support.rs`:
- `make_test_dir` (was duplicated in `scanner.rs` and
  `ebook::test_support`; the consolidated version uses the
  unique-per-invocation atomic counter so parallel runs don't collide)
- `seed_minimal_books` (was a private helper in `books/tests.rs`)
- `indexed` / `indexed_with_stat` (formerly `sync::test_helpers`)
- `seed_discovery_fixture`, `author_id_by_name`, `series_id_by_name`,
  `seed_books_for_one_author_and_series` (formerly
  `discovery::test_helpers`)
- `CoversTempDir` + `COVERS_ENV_LOCK` (formerly
  `covers::test_helpers`)

Files updated to import from `crate::test_support::*`:
- `db/src/author_photos_data.rs`
- `db/src/books/tests.rs`
- `db/src/browse.rs`
- `db/src/covers.rs`
- `db/src/discovery.rs`
- `db/src/ebook.rs` (keeps its module-local fixture helpers; re-exports
  `make_test_dir` so the existing `use crate::ebook::test_support::*;`
  globs in `ebook/{stat,parse,cover,accent}.rs` continue to resolve)
- `db/src/indexer.rs`
- `db/src/metadata_overrides.rs`
- `db/src/palette.rs`
- `db/src/scanner.rs`
- `db/src/settings.rs` (keeps its own `LibraryEnvGuard` — out of scope)
- `db/src/sync.rs`

`db/Cargo.toml` gains a `test-support` feature so the cfg gate compiles
for external test crates. Settings' `LibraryEnvGuard` stays local since
it's narrowly scoped to one module. Ebook-specific fixture loaders
(`solid_color_png`, `fixture`, `copy_fixture_into`,
`find_materialized_sidecar`) also stay put — they're internal to
ebook tests and listed nowhere else.

`cargo test -p omnibus-db` passes; `cargo clippy --all-targets
--all-features -- -D warnings` is clean.

Closes #306
