+++ db/src/author_photos_data.rs
@@ -205,11 +205,9 @@ pub async fn delete_author(pool: &SqlitePool, author_id: i64) -> Result<u64, sql
 mod tests {
     use super::*;
     use crate::books::list_books;
-    use crate::covers::test_helpers::CoversTempDir;
-    use crate::discovery::test_helpers::*;
     use crate::pool::init_db;
     use crate::sync::replace_books;
-    use crate::sync::test_helpers::indexed;
+    use crate::test_support::*;
 
     // -------------------------------------------------------------------------
     // F1.11 author profile photo tests
