+++ db/src/lib.rs
@@ -22,6 +22,8 @@ pub mod scanner;
 pub mod settings;
 pub mod sync;
 mod taxonomy;
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
 pub mod thumbs;
 pub mod worker;
